Changed transactional semantics to be more like fdb.
In Fowl's previous semantics, it was awkward to handle errors that occur in the execution of a transaction, because the transaction read operations would return promises that would not resolve until the developer already called .

This made it difficult to signal that a transaction failed halfway. With the new semantics it is easy to stop the commit, simply throw anywhere within the transaction body function or its Promise chain.